refactor(hooks): move `useForceUpdate` to `src`
