+++ packages/react-instantsearch-hooks/src/useForceUpdate.ts
